requirements.txt
aiofiles
aiolimiter
aiohttp
aioshutil
anytree
apscheduler
cloudscraper
dataclasses_json
dnspython
fastapi
feedparser
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
httpx
jinja2
lxml
m3u8
mutagen
natsort
par2cmdline-turbo
pathvalidate
Pillow
psutil
psycopg2-binary
pycryptodome
pycryptodomex
pymongo
pyrogram
python-dotenv
python-magic
python-qbittorrent
pysabnzbd
requests
telegraph
tenacity
tgcrypto
tidal-dl-ng
tidalapi
toml
typer
urllib3
uvicorn
uvloop
xattr
